flask: Converted /waypoints routes to proper blueprint
